+++ src/companies/index.ts
@@ -1,1 +1,2 @@
 export * from './companies.module';
+export * from './companies.service';
