+++ src/documents/contracts/index.ts
@@ -1,1 +1,2 @@
 export * from './contracts.module';
+export * from './contracts.service';
