+++ src/documents/invoice/index.ts
@@ -1,1 +1,2 @@
 export * from './invoice.module';
+export * from './invoice.service';
