+++ src/documents/payment/index.ts
@@ -1,1 +1,2 @@
 export * from './payment.module';
+export * from './payment.service';
