+++ src/warehouse/index.ts
@@ -1,1 +1,2 @@
 export * from './warehouse.module';
+export * from './warehouse.service';
